restructure for nuxt 4
