+++ collectivo/server/utils/zod.ts
@@ -1,0 +1,13 @@
+import { z } from "zod";
+
+export async function readDataorThrowError(event: any, schema: z.ZodTypeAny) {
+  const result = await readValidatedBody(event, (query) =>
+    schema.safeParse(query),
+  );
+
+  if (!result.success) {
+    throw result.error.issues;
+  }
+
+  return result.data;
+}
